create separate validation for register and login
