+++ client/js/screens/Register/styles.js
@@ -67,7 +67,7 @@ const styles = StyleSheet.create({
   },
   errorMsg: {
     color: "red",
-    marginTop: 20
+    marginTop: 10
   }
 });
